# AllAboard_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: AllAboard API
host: 'localhost'
schemes:
  - http
...
paths:
...
  /trainClass/trainClass:
    post:
      tags:
        - trainClass
      summary: Add a new train class type
      operationId: addTrainClassType
      consumes:
        - application/json
      produces: 
        - application/json
      parameters:
        - in: body
          name: trainClassType 
          description: Station object 
          required: true
          schema:
            $ref: '#/definitions/TrainClassType'
      responses:
        '200':
          description: Train Class Type created successfully
  /trainClass/trainClass/{classId}:
    put:
      tags:
        - trainClass
      summary: Update a train class type 
      operationId: updateTrainClassType
      parameters:
        - name: classId
          in: path 
          description: ID of the train class type to update 
          required: true
          type: integer
        - name: trainClassType
          in: body
          description: Updated train class type object 
          required: true
          schema: 
            $ref: '#/definitions/TrainClassType'
      responses:
        200:
          description: Train class type updated successfully
        400: 
          description: Invalid request data
        404:
          description: Train class type not found
    
    delete: 
      tags:
        - trainClass 
      summary: Delete a train class type 
      operationId: deleteTrainClassType
      parameters:
        - name: classId
          in: path
          description: ID of the train class type to delete
          required: true
          type: integer
      responses:
        200:
          description: Train class type deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Train class type not found
...
  /train/train:
    post:
      tags:
        - train 
      summary: Add a new train 
      operationId: addTrain
      consumes: 
        - application/json
      produces: 
        - application/json 
      parameters: 
        - in: body 
          name: train 
          description: Train object 
          required: true 
          schema: 
            $ref: '#/definitions/Train'
      responses: 
        '200': 
          description: Train created successfully 
  /train/train/{trainId}:
    put:
      tags:
        - train
      summary: Update a train
      operationId: updateTrain
      parameters:
        - name: trainId
          in: path 
          description: ID of the train to update 
          required: true
          type: integer
        - name: train
          in: body
          description: Updated train 
          required: true
          schema: 
            $ref: '#/definitions/Train'
      responses:
        200:
          description: Train updated successfully
        400: 
          description: Invalid request data
        404:
          description: Train not found
    
    delete:
      tags:
        - train 
      summary: Delete a train 
      operationId: deleteTrain
      parameters:
        - name: trainId
          in: path
          description: ID of the train to delete
          required: true
          type: integer
      responses:
        200:
          description: Train deleted successfully
        400:
          description: Invalid request data 
        404:
          description: Train not found  
  /train/getTrainsBasedOnStation:
    get:
      tags:
        - train
      summary: Get trains based on station
      operationId: getTrainsBasedOnStation
      parameters:
        - name: from
          in: query
          description: The source station
          required: true
          type: string
        - name: to
          in: query
          description: The destination station
          required: true
          type: string
        - name: date
          in: query
          description: The travel date
          required: true
          type: string
          format: date
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Train'
...
definitions:
...
  TrainClassType:
    type: object
    properties:
      train_class_id:
        type: integer
      train_class_name:
        type: string
...
  TrainClass:
    type: object 
    properties:
      train_class_type:
        type: string 
      capacity:
        type: integer
      base_price: 
        type: number 
  TrainSeat: 
    type: object 
    properties: 
      seat_type: 
        type: string 
      class_type:
        type: string
      seat_number:
        type: string
      is_booked: 
        type: boolean 
      seat_price:
        type: number
  Train:
    type: object
    properties:
      train_id:
        type: integer
      train_name:
        type: string
      source_station:
        type: string
      destination_station:
        type: string
      travel_date:
        type: string
        format: date
      departure_time:
        type: string
        format: time
      train_classes:
        type: array
        items:
          $ref: '#/definitions/TrainClass'
      train_seats: 
        type: array 
        items: 
          $ref: '#/definitions/TrainSeat'
...